+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 _Copyright (c) 2013 Daniele Veneroni._  
 _Released under GPLv3 License. See [LICENSE.md](LICENSE.md) for further information._
 
+### Update 29/11/2013
+
+* Changed graphic engine from Raphael.js (SVG) to Palette.js (Canvas)
+* Improved graphic performance (thanks to canvas and Palette.js)
+* Now full spectrum is displayed (before it was trunked)
+
 ### Update 28/11/2013 (2)
 
 * New Loading Message